requirements.txt
asgiref==3.6.0
Django==4.1.7
django-adminactions==2.0.0
psycopg2-binary==2.9.5
python-decouple==3.8
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
whitenoise==6.4.0
xlrd==2.0.1
xlwt==1.3.0
